Consolidate options into client
